# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MLOVI - Multidisciplinar: BACKEND API V1",
        "description": "Description of your API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Main (production) server"
        }
    ],
    "paths": {
...
        "/guias": {
            "get": {
                "tags": [
                    "Guias"
                ],
                "summary": "Listar todos as guias",
                "operationId": "getGuiaList",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    }
                }
            },
            "post": {
                "tags": [
                    "Guias"
                ],
                "summary": "Cadastrar uma nova guia",
                "description": "Cria um novo registro de guia com os dados fornecidos.",
                "operationId": "storeGuia",
                "requestBody": {
                    "description": "Dados da guia",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lote_numero",
                                    "id_convenio",
                                    "id_paciente",
                                    "id_procedimento",
                                    "status",
                                    "data",
                                    "quantidade",
                                    "valor",
                                    "id_contratado_executante",
                                    "id_profissional_executante"
                                ],
                                "properties": {
                                    "lote_numero": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_convenio": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_paciente": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_procedimento": {
                                        "type": "intenger",
                                        "example": "14"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": 0
                                    },
                                    "num_guia_prestador": {
                                        "type": "string",
                                        "example": "GP001"
                                    },
                                    "num_guia_operadora": {
                                        "type": "string",
                                        "example": "GO001"
                                    },
                                    "data": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-29"
                                    },
                                    "quantidade": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "valor": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "tabela": {
                                        "type": "string",
                                        "example": "00"
                                    },
                                    "id_contratado_executante": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "codigo_operadora_contratado": {
                                        "type": "string",
                                        "example": "COD001"
                                    },
                                    "cnpj_contratado": {
                                        "type": "string",
                                        "example": "12345678901234"
                                    },
                                    "cnes_contratado": {
                                        "type": "string",
                                        "example": "CNES001"
                                    },
                                    "id_profissional_executante": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "codigo_operadora_profissional": {
                                        "type": "string",
                                        "example": "COD002"
                                    },
                                    "id_conselho": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "numero_conselho_profissional": {
                                        "type": "string",
                                        "example": "08"
                                    },
                                    "uf_conselho_profissional": {
                                        "type": "string",
                                        "example": 26
                                    },
                                    "cbo_s": {
                                        "type": "string",
                                        "example": "251510"
                                    },
                                    "tipo_consulta": {
                                        "type": "string",
                                        "example": "4"
                                    },
                                    "indicador_acidente": {
                                        "type": "string",
                                        "example": "9"
                                    },
                                    "cobertura_especial": {
                                        "type": "string",
                                        "example": "02"
                                    },
                                    "atendimento_regime": {
                                        "type": "string",
                                        "example": "01"
                                    },
                                    "observacao": {
                                        "type": "string",
                                        "example": "Observações adicionais"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guia criado com sucesso"
                    },
                    "400": {
                        "description": "Erro na requisição"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/guias/{id}": {
            "get": {
                "tags": [
                    "Guias"
                ],
                "summary": "Buscar guia por ID",
                "operationId": "getGuiaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do guia",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Guia não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Guias"
                ],
                "summary": "Atualizar guia",
                "operationId": "updateGuia",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do guia",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do guia a serem alterados. Verifique os campos no Schema",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Guia Ilha do leite"
                                    },
                                    "endereco": {
                                        "type": "string",
                                        "example": "Av. Frei Matias Teves, 280 - Ilha do Leite, Recife - PE, 50070-465"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Guia não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Guias"
                ],
                "summary": "Excluir guia",
                "operationId": "deleteGuia",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do guia",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Guia excluído com sucesso"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Guia não encontrado"
                    }
                }
            }
        },
...
    },
...
}